improved FireShader | added noise Textures

+++ Assets/Spells/Shader/ShieldTransparent.shader
@@ -60,7 +60,7 @@
 				v2f o;
 				o.vertex= UnityObjectToClipPos(v.vertex);
 				o.posWorld = mul(unity_ObjectToWorld, v.vertex);
-				o.normalDir = normalize(mul(float4(v.normal, 0.0), unity_WorldToObject).xyz);
+				o.normalDir = normalize(mul(half4(v.normal, 0.0), unity_WorldToObject).xyz);
 				o.uv = UnityObjectToClipPos(v.vertex);
 				o.alpha = v.uv;
